Add files via upload
Current working versions of the team building. Original algorithm works in the command line and is finicky on spelling and capitalization. 

Please refer to versions 1.0 and 1.1 which operate as a separate GUI that takes two pokemon's type input from drop down menus and outputs a suggested third pokemon. The two versions operate the same. Version 1.1 is more efficient and removes a barrier on how many type weaknesses two pokemon can collectively get while allowing the program to run.